requirements.txt
Django==1.9.4
requests==2.2.1
pillow==3.1.1
markdown==2.6.6
django-filter==0.13
django-crispy-forms==1.6.0
django-guardian==1.4.2
djangorestframework==3.3.3